Update NeuTrace2.0.ijm
Fixed minor bugs. Changed the way ImageJ saves the combined plot at the end so that now it can be re-opened in ImageJ and edited as a plot. Added calculations to the process so that the macro computes the pixel-by-pixel ratio of the two channels, then gets statistics on these values for a give neurite trace, then saves them in the metadata file.